requirements.txt
discord.py
groq
python-dotenv
qdrant-client
pydantic
sentence-transformers
